pyproject.toml
[tool.poetry]
name = "tiger"
version = "0.1.0"
description = ""
authors = ["Erick <efm.physics@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.13"
streamlit = "^1.23.1"
streamlit-authenticator = "^0.2.2"
stauth = "^0.1.9"
google-cloud-bigquery = "^3.11.0"
db-dtypes = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
